Got halfway into this project, part of London App Brewery's 2022 Web Development Bootcamp online course, when I realized I forgot to start a repo.
The HTML and CSS are mostly provided by the course.  The contents of index.js are, painfully, hand-written by me, about six times because I kept mistyping *something* by trying to cheat using copy-and-paste, so I finally gave in and hand typed everything.

This is a great way to learn to code, if, as I said, painful. :)

CURRENTLY the drum kit has sounds working when clicking.  See comments in file for further details and TODO is to come next commit :)